pyproject.toml
[project]
name = "kartoza-django"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Riski Z", email = "riski.zulfiansyah@gmail.com" }
]
dependencies = [
    "django>=5.1.7",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.10",
    "django-leaflet>=0.31.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/kartoza_django"]
